requirements.txt
maafw>=4.0.0
loguru
pandas
openpyxl
opencv-python
zhconv